fix : self like can't
- 본인 글, 댓글은 좋아요를 누를 수 없도록 수정

+++ src/main/java/com/fiveman/newsfeed/comment/dto/CommentResponseDto.java
@@ -15,7 +15,7 @@ public class CommentResponseDto {
     private Long userid;
     private Long commentId;
     private String content;
-    private Long LikeCount;
+    private Integer LikeCount;
     private LocalDateTime createAt;
     private LocalDateTime updatedAt;
